# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger representation of the application's API",
        "title": "Donkey API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "vk@cloudnation.dk"
        },
        "license": {
            "name": "MIT",
            "url": "./LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "get accounts",
                "tags": [
                    "accounts"
                ],
                "summary": "Login with your username and password",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Authenticated\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Return a user information",
                "tags": [
                    "accounts"
                ],
                "summary": "Fetch a user information",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Authenticated\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}